Implement SafeString escaping passthrough
Handlebars.js and other templating engines offer a way to return a value
from a helper and not have it escaped based on the helpers choice, rather
than the templates.  This adds a new feature flag and associated code to
create a new LCSafeString class.  When returned from a helper the contents
of the LCSafeString are output as-is.

There may be some interactions with LCRun2::raw that i don't understand yet.

Addresses https://github.com/zordius/lightncandy/issues/25